refactor(docs): extract mdx component map
